Extract L4A GEDI above ground biomass data for the GEDI AGB prep function.

Extract L4A GEDI above ground biomass data for the GEDI AGB prep function.

GEDI_AGB_extract(
  site_info,
  start_date,
  end_date,
  outdir,
  nfile.min = 0,
  nrow.min = 0,
  buffer = 0.01,
  gradient = 0
)

Arguments

site_info

A list including site_id, longitude, and latitude.

start_date

target start date (date with YYYY-MM-DD format) for preparing GEDI AGB from remote or local database.

end_date

end date (date with YYYY-MM-DD format) for preparing GEDI AGB from remote or local database.

outdir

Directory where the final CSV file will be stored.

nfile.min

the minimum required file number to be downloaded and extracted, default is 0.

nrow.min

the minimum required observation number to be extracted, default is 0.

buffer

buffer distance (in degrees) for locate GEDI AGB searching box (default is 0.01 [~ 1 km]).

gradient

the gradient for iteratively enlarge the extent if the nfile.min or nrow.min are not reached, default is 0. If nfile.min or nrow.min is 0 this will be skipped.

Value

A list of AGB data frames for each site.

Author

Dongchen Zhang