Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.

Prepare MODIS LAI data from the NASA DAAC server for the SDA workflow.

Usage

MODIS_LAI_ts_filter(lai.csv, boundary = c(0.05, 0.95))

Arguments

lai.csv

data frame: A data frame containing date, site id, lat, lon, lai, sd, and qc columns.

boundary

numeric: A vector contains the upper and lower quantiles for filtering lai records. The default is c(0.05, 0.95).

Value

A data frame containing date, site id, lat, lon, lai, sd, and qc columns.

Author

Dongchen Zhang