Create URL that can be used to request data from NASA DAAC server.

Create URL that can be used to request data from NASA DAAC server.

NASA_DAAC_URL(
  base_url = "https://cmr.earthdata.nasa.gov/search/granules.json?pretty=true",
  provider,
  page_size = 2000,
  page = 1,
  concept_id,
  bbox,
  daterange = NULL
)

Arguments

base_url

Character: base URL for the CMR search. default is "https://cmr.earthdata.nasa.gov/search/granules.json?pretty=true".

provider

Character: ID of data provider from NASA DAAC. See `NASA_CMR_finder` for more details.

page_size

Numeric: maximum requested length, default is 2000.

page

Numeric: which page of the URL, default is 1.

concept_id

Character: CMR Concept ID. See `NASA_CMR_finder` for more details.

bbox

Numeric: vector of bounding box coordinates.

daterange

Character: vectors of the requested start and end dates. In the form "yyyy-mm-dd".

Value

A character of URL that can be used to request data.

Author

Dongchen Zhang

Examples

if (FALSE) {
provider <- "ORNL_CLOUD"
concept_id <- "C2770099044-ORNL_CLOUD"
bbox <- "-121,33,-117,35"
daterange <- c("2022-02-23", "2022-05-30")
URL <- NASA_DAAC_URL(provider = provider, 
concept_id = concept_id, 
bbox = bbox, 
daterange = daterange)
}