Prepare SMAP soil moisture profile (SMP) data from the NASA DAAC server for the SDA workflow. The CSV file that works for the `SMAP_SMP_prep` function will be exported.

Prepare SMAP soil moisture profile (SMP) data from the NASA DAAC server for the SDA workflow. The CSV file that works for the `SMAP_SMP_prep` function will be exported.

Usage

Prep.SMAP.CSV.from.DAAC(
  site_info,
  extent,
  from,
  to,
  download.outdir,
  csv.outdir,
  credential_path
)

Arguments

site_info

list: Bety list of site info including site_id, lon, and lat.

extent

numeric: A vector contains the bounding box that covers all sites (West longitude, East longitude, South latitude ,North latitude).

from

character: the start time for searching the MODIS products.

to

character: the end time for searching the MODIS products.

download.outdir

character: Where the MODIS tiles will be stored.

csv.outdir

character: Where the final CSV file will be stored.

credential_path

Character: physical path to the credential file (.netrc file).

Value

A data frame containing SMP and sd for each site and each time step.

Author

Dongchen Zhang