Prepare SMAP Soil Moisture (SMP) data for the SDA workflow.

Prepare SMAP Soil Moisture (SMP) data for the SDA workflow.

Usage

SMAP_SMP_prep(
  site_info,
  start_date,
  end_date,
  time_points,
  outdir,
  search_window = 30,
  export_csv = TRUE,
  update_csv = FALSE
)

Arguments

site_info

Bety list of site info including site_id, lon, and lat.

start_date

Start date of SDA workflow.

end_date

End date of SDA workflow.

time_points

A vector contains each time point within the start and end date.

outdir

Where the final CSV file, and the CSV file from GEE are stored.

search_window

search window for locate available SMP values.

export_csv

Decide if we want to export the CSV file.

update_csv

Decide if we want to update current CSV file given an updated SMAP_gee.csv file

Value

A data frame containing SMAP smp and sd for each site and each time step.

Author

Dongchen Zhang