download.NLCD

Downloads and unzips the National Land Cover Database http://www.mrlc.gov/nlcd2011.php. Will automatically insert into PEcAn database if database connection provided.

download.NLCD(outdir, year = 2011, con = NULL)

Arguments

outdir

Directory to download NLCD to

year

which NLCD year to download. Only 2001 and 2011 are currently supported.

con

Optional database connection. If specified then the code will check to see if the file already exists in PEcAn before downloading, and will also create a database entry for new downloads

Author

Mike Dietze