extract.LandTrendr.AGB

extract.LandTrendr.AGB

extract.LandTrendr.AGB(
  site_info,
  dataset = "median",
  buffer = NULL,
  fun = "mean",
  data_dir = NULL,
  product_dates = NULL,
  output_file = NULL,
  ...
)

Arguments

site_info

list of site info for parsing AGB data: list(site_id, site_name, lat, lon, time_zone)

dataset

Which LandTrendr dataset to parse, "median" or "stdv".Default: "median"

buffer

Optional. operate over desired buffer area (not yet implemented)

fun

Optional function to apply to buffer area. Default - mean

data_dir

directory where input data is located. Can be NUL if con is specified

product_dates

Process and extract data only from selected years. Default behavior (product_dates = NULL) is to extract data from all availible years in BETYdb or data_dir

output_file

Path to save LandTrendr_AGB_output.RData file containing the output extraction list (see return)

...

Additional arguments, currently ignored

Value

list of two containing the median AGB values per pixel and the corresponding standard deviation values (uncertainties)

Author

Shawn Serbin, Alexey Shiklomanov

Examples

if (FALSE) {

# Example 1 - using BETYdb site IDs to extract data
# Database connection (optional)

con <- PEcAn.DB::db.open(
  list(user='bety', password='bety', host='localhost',
  dbname='bety', driver='PostgreSQL',write=TRUE))

site_ID <- c(2000000023,1000025731,676,1000005149) # BETYdb site IDs
suppressWarnings(site_qry <- glue::glue_sql("SELECT *, ST_X(ST_CENTROID(geometry)) AS lon, 
ST_Y(ST_CENTROID(geometry)) AS lat FROM sites WHERE id IN ({ids*})", 
ids = site_ID, .con = con))
suppressWarnings(qry_results <- DBI::dbSendQuery(con,site_qry))
suppressWarnings(qry_results <- DBI::dbFetch(qry_results))
site_info <- list(site_id=qry_results$id, site_name=qry_results$sitename, lat=qry_results$lat, 
lon=qry_results$lon, time_zone=qry_results$time_zone)
data_dir <- "~/scratch/agb_data/"

results <- extract.LandTrendr.AGB(site_info, "median", buffer = NULL, fun = "mean", 
data_dir, product_dates, output_file)

}