Merge image tiles to a single image (currently support hdf and tif image format).

Merge image tiles to a single image (currently support hdf and tif image format).

Usage

merge_image_tiles(
  in.path,
  out.path = NULL,
  band.name = NULL,
  just.band.name = TRUE,
  keep.files = FALSE,
  skip.conversion = FALSE,
  image.settings = list(crs = "EPSG:4326", dimension = NULL, ext = NULL, fun = NULL),
  computation = list(GDAL_CACHEMAX = 1000, wm = "80%", NUM_THREADS =
    parallel::detectCores() - 1, COMPRESS = "DEFLATE")
)

Arguments

in.path

character: physical path to the folder that contains all the original image tiles.

out.path

character: physical path to the folder that contains converted and merged images.

band.name

character: band name of the image. Default is NULL.

just.band.name

logical: if we just want the band names of the image file. Default is TRUE.

keep.files

logical: if we want to keep the image tiles at the end.

skip.conversion

logical: if we want to ignore the image conversion. Note that this is a experimental feature, which only works when images are all in the GeoTIFF format.

image.settings

list: settings used during exporting merged image. Such as image coordinate system (crs), dimension, extents (ext), and average function (fun).

computation

list: settings used for configuring computation. Such as maximum memory per CPU (GDAL_CACHEMAX), percentage of total memory (wm), number of CPUs (NUM_THREADS), compress method (COMPRESS).

Value

character: file path to the merged GeoTIFF file.

Details

Please make sure all image tiles are stored in the `folder.path`. Please refer to the gdalwarp manual for more details https://gdal.org/en/stable/programs/gdalwarp.html

Author

Dongchen Zhang