Insert the output data returned from rp_control into BETYdb

Insert the output data returned from rp_control into BETYdb

remotedata_db_insert(
  output,
  remotefile_check_flag,
  siteid,
  out_get_data,
  out_process_data,
  write_raw_start,
  write_raw_end,
  write_pro_start,
  write_pro_end,
  raw_check,
  pro_check,
  raw_mimetype,
  raw_formatname,
  pro_mimetype,
  pro_formatname,
  dbcon
)

Arguments

output

output list from rp_control

remotefile_check_flag

remotefile_check_flag

siteid

siteid

out_get_data

out_get_data

out_process_data

out_process_data

write_raw_start

write_raw_start, start date of the raw file

write_raw_end

write_raw_end, end date of the raw file

write_pro_start

write_pro_start

write_pro_end

write_pro_end

raw_check

id, site_id, name, start_date, end_date, of the existing raw file from inputs table and file_path from dbfiles tables

pro_check

pro_check id, site_id, name, start_date, end_date, of the existing processed file from inputs table and file_path from dbfiles tables

raw_mimetype

raw_mimetype

raw_formatname

raw_formatname

pro_mimetype

pro_mimetype

pro_formatname

pro_formatname

dbcon

BETYdb con

Value

list containing raw_id, raw_path, pro_id, pro_path

Author

Ayush Prasad

Examples